Add mostRecentAccessTime attribute to streams
This commit adds a mostRecentAccessTime attribute to all streams.  For
input, this corresponds to the time for which input was last
read.  For output, this corresponds to the time for which output was
last written.  Note that for output, there is an option to specify an
output time that is not the current time; the value used here for
mostRecentAccessTime uses the time actually sent to the file.

This attribute is initialized to '9999-12-31_23:59:59' and will contain
that value prior to the first access.